[CHORE] api 요청 코드 파일 정리
