+++ client/src/hooks/WalksQuery.ts
@@ -3,6 +3,24 @@ import { useQuery } from 'react-query';
 import { API } from '../apis/api';
 import { WalkDetail } from '../models/WalkDefault';
 
+export function useGetWalksQuery(query: string) {
+  let url: string;
+
+  if (query) {
+    url = `${API.WALKS}${query}`;
+  } else {
+    url = API.WALKS;
+  }
+
+  return useQuery(
+    'walks',
+    async () =>
+      await axios.get(url).then(async (res) => {
+        return res.data.communityList;
+      })
+  );
+}
+
 export function useWalksDetailQuery(communityId: string) {
   const { data, error } = useQuery('communityDetail', async () => {
     const { data } = await axios.get<WalkDetail>(`${API.WALKS}/${communityId}`);
