+++ client/src/pages/walks/[walkId]/notice.tsx
@@ -2,7 +2,7 @@ import { GetServerSideProps } from 'next';
 import TabTitle from '../../../components/TabTitle';
 import DetailLayout from '../../../components/walks/walksDetail/DetailLayout';
 import MoimNotice from '../../../components/walks/walksDetail/notice/MoimNotice';
-import { useWalksDetailQuery } from '../../../hooks/WalksDetailQuery';
+import { useWalksDetailQuery } from '../../../hooks/WalksQuery';
 
 export default function Notice({ walkId }: { walkId: string }) {
   const walkDetail = useWalksDetailQuery(walkId);
